+++ packages/core/client/Place.js
@@ -78,9 +78,9 @@ Kepler.Place = Class.extend({
 			var iconOpts = K.settings.public.map.icon;
 			
 			self.icon = new L.NodeIcon({
-				/*iconSize: new L.Point(iconOpts.iconSize),
+				/*conSize: new L.Point(iconOpts.iconSize),
 				iconAnchor: new L.Point(iconOpts.iconAnchor),
-				popupAnchor: new L.Point(iconOpts.popupAnchor),*/
+				popupAnchor: new L.Point(iconOpts.popupAnchor),*/				
 				nodeHtml: L.DomUtil.create('div')
 			});