+++ packages/core/modules/templates.js
@@ -13,6 +13,7 @@ Kepler.templates = {
 	popupUser: [],
 	popupCursor: [],
 	popupGeojson: [],
+	markerCursor: [],
 	markerPlace: [],	
 	markerUser: [],
 	itemPlace:  [],