+++ packages/edit/package.js
@@ -11,7 +11,7 @@ Package.onUse(function(api) {
   api.versionsFrom("1.5.1");
 
   api.use([
-    'keplerjs:core@1.2.5',
+    'keplerjs:core',
   ]);
 
   api.addFiles([