+++ packages/ui/client/views/markers.js
@@ -1,0 +1,14 @@
+
+
+/*Template.markerPlace.helpers({
+	checkinsCount: function(e) {
+		return 3;
+	}
+});
+*/
+
+/*Template.markerCluster.helpers({
+	checkinsCount: function(e) {
+		return 3;
+	}
+});*/