+++ packages/ui/package.js
@@ -18,7 +18,7 @@ Package.onUse(function(api) {
   api.versionsFrom("1.5.1");
   
   api.use([
-    'keplerjs:core@1.2.5',
+    'keplerjs:core',
   ]);
 
   api.addFiles(globsync('client/**/*.*'), 'client');