+++ packages/admin/package.js
@@ -2,7 +2,7 @@
 Package.describe({
   name: 'keplerjs:admin',
   summary: 'keplerjs plugin for administration of platform',
-  version: "1.2.5",
+  version: "1.3.0",
   git: "https://github.com/Keplerjs/Kepler.git"
 });