+++ packages/edit/package.js
@@ -2,7 +2,7 @@
 Package.describe({
   name: 'keplerjs:edit',
   summary: 'keplerjs plugin for edit place informations',
-  version: "1.2.5",
+  version: "1.3.0",
   git: "https://github.com/Keplerjs/Kepler.git"
 });