+++ packages/favorites/package.js
@@ -2,7 +2,7 @@
 Package.describe({
   name: 'keplerjs:favorites',
   summary: 'keplerjs plugin favorites places',
-  version: "1.2.5",
+  version: "1.3.0",
   git: "https://github.com/Keplerjs/Kepler.git"
 });