+++ packages/osm/plugin.js
@@ -1,17 +1,8 @@
-/*
-
-//TODO share obejcts to overpass by osm node or way id
-
-[out:json];way(42764800);(._;>;);
-out body;way(56219784);(._;>;);
-out body;
-
-*/
 
 K.Plugin({
 	name: 'osm',
 	templates: {
-		panelPlace: ['panelPlace_osm'],
+		panelPlace: 'panelPlace_osm',
 		popupCursor: 'popupCursor_osm'
 	},
 	filters: {