+++ packages/routing/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: "keplerjs:routing",
   summary: "Keplerjs routing plugin",
-  version: "1.2.4",
+  version: "1.3.0",
   git: "https://github.com/Keplerjs/Kepler.git"
 });