+++ packages/share/plugin.js
@@ -2,7 +2,10 @@
 K.Plugin({
 	name: 'share',
 	templates: {
-		panelPlace: { name:'panelPlace_share', order: -5 },
+		panelPlace: {
+			'panelPlace_share': {order: -5 },
+			//'popupCursor_share': {order: -6 }
+		},
 		popupCursor: 'popupCursor_share',
 		//popupUser: 'popupCursor_share'
 	}