+++ packages/ui/client/views/pluginsTemplate.js
@@ -8,6 +8,8 @@
 
 Template.pluginsTemplate.helpers({
 	templates: function() {
+		if(!Template[this.name]) return [];
+
 		return K.Plugin.templatesByPlaceholder(this.name, Template.parentData());
 	}
 });