+++ CHANGELOG.md
@@ -4,9 +4,13 @@ To see all releases [here](https://github.com/Keplerjs/Kepler/releases)
 ## Next version
 - Meteor 1.6.1 compatibility
 - update npm deps
-- support templates custom ordering
-- rewrite Place icon building using templates
-- defined new templates placeholders: markerPlace, markerUser, markerCursor
+
+## 1.3.0
+- changes *K.templates* definition format
+- support templates plugin custom ordering
+- override templates config by *settings.json*
+- rewrite Place icon building using Meteor templates
+- defined new templates placeholders: *markerPlace*, *markerUser*, *markerCursor*
 
 ## v1.2.5
 - plugins renamed dirs in plain name